+++ CHANGES.md
@@ -1,4 +1,6 @@
 
+  - added `tanimoto` (Tanimoto score) (contribution by François Berenger)
+
 # 1.6 (06/02/2021)
   - fixed build on 32-bit platforms e.g. js_of_ocaml (patch from Tim Bourke)
   - added `exceeds_max_length` to replace `max_length` (deprecated)